Added delay_us and use_hrtimer module params.
Observation, the DMA callback gets called when data has been transferred
to the McBSP not when the McBSP is done pumping out the data.

Therefore delay calculations need to compensate for this since we are
initiating the next transfer from the DMA callback.

I'll look into a McBSP interrupt for transfer complete, but the calculation
to compensate is not difficult.

I also removed most of the printks. Things are working well enough they
aren't needed.